you can damage them now

+++ src/game/ship.py
@@ -48,7 +48,7 @@ class Ship():
     def draw(self, cam: camera.Camera):
         map.MapRenderer.draw_map(cam, pygame.Vector2(), self.map_data, consts.DRAW_SCALE)
 
-    def get_tile_center(self, cam: camera.Camera, x: int, y: int) -> pygame.Vector2:
+    def get_tile_center(self, cam: camera.Camera, x: float, y: float) -> pygame.Vector2:
         return map.MapRenderer.get_tile_center(
             x, y, cam, pygame.Vector2(), self.map_data, consts.DRAW_SCALE
         )
